requirements.txt
africastalking==1.2.8
asgiref==3.8.1
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.1
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==29.0.0
idna==3.10
iniconfig==2.0.0
josepy==1.14.0
mozilla-django-oidc==4.0.1
packaging==24.1
phonenumbers==8.13.45
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
schema==0.7.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
